# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Senac Bike Store API - By Condé",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "INSIRA AQUI A URL do Host",
            "description": "Produção"
        },
        {
            "url": "INSIRA AQUI A URL do Host",
            "description": "Produção"
        }
    ],
    "paths": {
...
        "/api/marcas": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Retorna a Lista de Marcas.",
                "description": "Retorna o JSON da Lista de Marcas.",
                "operationId": "getMarcasList",
                "parameters": [
                    {
                        "name": "ordenacao",
                        "in": "query",
                        "description": "Ordernação dos dados",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcasResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcasResponseException500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Armazena os dados de uma nova Marca.",
                "description": "Armazena e retorna o JSON de uma nova Marca.",
                "operationId": "postMarca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarcaNewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException500"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/marcas/{id}": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Retorna os dados de uma Marca.",
                "description": "Retorna o JSON de uma Marca.",
                "operationId": "getMarca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Apaga os dados de uma categoria.",
                "description": "Apaga e retorna o JSON de confirmação.",
                "operationId": "deleteMarca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Atualiza os dados de uma marca.",
                "description": "Atualiza e retorna o JSON de uma nova Marca.",
                "operationId": "patchMarca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Marca"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MarcaNewRequest": {
                "title": "MarcaNewRequest",
                "description": "Requisição enviada via Body para uma inserção ou atualização de uma Marca",
                "required": [
                    "nome_da_Marca"
                ],
                "properties": {
                    "nome_da_marca": {
                        "title": "nome_da_Marca",
                        "description": "Nome da Marca.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Marca": {
                "title": "Marca",
                "description": "Modelo da Marca.",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Id da Marca.",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nome_da_marca": {
                        "title": "nome da marca",
                        "description": "Nome da Marca.",
                        "type": "string",
                        "example": "Electra"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Marca"
                }
            },
...
            "MarcaResponse": {
                "title": "MarcaResponse",
                "description": "Resposta enviada via JSON para os dados de uma Marca",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Marca retornada."
                    },
                    "marca": {
                        "$ref": "#/components/schemas/Marca"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MarcaResponse"
                }
            },
...
            "MarcaResponseException406": {
                "title": "MarcaResponseException406",
                "description": "Resposta enviada via JSON para os dados de uma Marca",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 406
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "O campo Id deve ser numérico."
                    },
                    "marca": {
                        "title": "Marca",
                        "description": "Dados de uma Marca.",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MarcaResponseException406"
                }
            },
            "MarcaResponseException500": {
                "title": "MarcaResponseException500",
                "description": "Resposta enviada via JSON para os dados de uma Marca",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 500
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Erro Interno."
                    },
                    "marca": {
                        "title": "Marca",
                        "description": "Dados de uma Marca.",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MarcaResponseException500"
                }
            },
            "MarcasResponse": {
                "title": "MarcasResponse",
                "description": "Resposta enviada via JSON para Lista de Marcas",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Lista de Marcas retornada."
                    },
                    "Marcas": {
                        "title": "Marcas",
                        "description": "Lista de Marcas retornada.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Marca"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MarcasResponse"
                }
            },
            "MarcasResponseException500": {
                "title": "MarcasResponseException500",
                "description": "Resposta enviada via JSON para Lista de Marcas",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 500
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Erro Interno."
                    },
                    "Marcas": {
                        "title": "Marcas",
                        "description": "Lista de Marcas retornada.",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MarcasResponseException500"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    }
}